Warning fix for wxOS2 night build.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@38417 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/os2/listbox.cpp
@@ -607,7 +607,7 @@ void wxListBox::SetString(unsigned int n, const wxString& rsString)
 
     int nNewN = n;
 
-    if (n == (int)(m_nNumItems - 1))
+    if (n == (m_nNumItems - 1))
         nNewN = -1;
 
     ::WinSendMsg( GetHwnd()
